+++ src/flows/ContractorOnboarding/api.ts
@@ -15,8 +15,9 @@ import { selectContractorSubscriptionStepSchema } from '@/src/flows/ContractorOn
 import {
   JSONSchemaFormResultWithFieldsets,
   FlowOptions,
+  JSFModify,
 } from '@/src/flows/types';
-import { formatCurrency } from '@/src/lib/utils';
+import { clearBase64Data, formatCurrency } from '@/src/lib/utils';
 import { Client } from '@/src/client/client';
 import { createHeadlessForm } from '@/src/common/createHeadlessForm';
 import { useMutation, useQuery, UseQueryResult } from '@tanstack/react-query';
@@ -35,12 +36,18 @@ export const useGetContractDocumentSignatureSchema = ({
   options,
 }: {
   fieldValues: FieldValues;
-  options?: { queryOptions?: { enabled?: boolean } };
+  options?: { queryOptions?: { enabled?: boolean }; jsfModify?: JSFModify };
 }) => {
   return useQuery({
-    queryKey: ['contract-document-signature'],
+    queryKey: [
+      'contract-document-signature',
+      fieldValues.review_completed,
+      options?.jsfModify,
+    ],
     queryFn: async () => {
-      return createHeadlessForm(signatureSchema, fieldValues);
+      return createHeadlessForm(signatureSchema, fieldValues, {
+        jsfModify: options?.jsfModify,
+      });
     },
     enabled: options?.queryOptions?.enabled,
   });
@@ -90,7 +97,7 @@ export const useGetShowContractDocument = ({
 }: {
   employmentId: string;
   contractDocumentId: string;
-  options?: { queryOptions?: { enabled?: boolean } };
+  options?: { queryOptions?: { enabled?: boolean }; jsfModify?: JSFModify };
 }) => {
   const { client } = useClient();
   return useQuery({
@@ -103,7 +110,15 @@ export const useGetShowContractDocument = ({
     },
     enabled: options?.queryOptions?.enabled,
     select: ({ data }) => {
-      return data?.data;
+      return {
+        ...data?.data,
+        contract_document: {
+          ...data?.data?.contract_document,
+          content: clearBase64Data(
+            data?.data?.contract_document?.content as $TSFixMe,
+          ),
+        },
+      };
     },
   });
 };