+++ src/flows/Termination/components/TerminationDialogInfoContent/TerminationDialogInfoContent.tsx
@@ -4,14 +4,14 @@ import { cn } from '@/src/internals';
 
 const terminationSteps = [
   'You submit your termination request.',
-  'We review your request along with the employment agreement, local labor laws, and any collective bargaining agreements in place.',
+  'Remote reviews your request along with the employment agreement, local labor laws, and any collective bargaining agreements in place.',
   'A lifecycle specialist contacts you if they need additional details. Sometimes they will request a call.',
-  'Depending on the country-specific laws and the terms of the employment agreement, we may need to seek additional legal advice. We’ll let you know if there are additional charges before proceeding.',
+  'Depending on the country-specific laws and the terms of the employment agreement, Remote may need to seek additional legal advice. We’ll let you know if there are additional charges before proceeding.',
   'Terms and conditions for the termination are finalized.',
   'You verbally communicate the termination to the employee.',
-  'We share termination documents with the employee.',
+  'Remote shares termination documents with the employee.',
   'The employee signs the termination documents if required.',
-  'We process the final salary and offboard the employee.',
+  'Remote processes the final salary and offboards the employee.',
 ];
 
 export const TerminationDialogInfoContent = () => {
@@ -23,8 +23,8 @@ export const TerminationDialogInfoContent = () => {
       <p
         className={cn('RemoteFlows__TerminationDialogInfoContent__Description')}
       >
-        Please do not inform the employee of the termination. We'll lead
-        communication with them after you submit this form and we finalize the
+        Please do not inform the employee of the termination. Remote will lead
+        communication with them after you submit this form and they finalize the
         termination terms.
       </p>
       <p
@@ -32,8 +32,8 @@ export const TerminationDialogInfoContent = () => {
           'RemoteFlows__TerminationDialogInfoContent__Description RemoteFlows__TerminationDialogInfoContent__FollowUpDescription',
         )}
       >
-        If you already informed them, that's okay. We'll ask for the details
-        regarding that communication and take it from there.
+        If you already informed them, that's okay. Remote will ask for the
+        details regarding that communication and they'll take it from there.
       </p>
 
       <ZendeskTriggerButton