+++ src/index.tsx
@@ -74,6 +74,8 @@ export type {
   ContractPreviewResponse,
 } from '@/src/flows/ContractorOnboarding';
 
+export type { ContractPreviewStatementProps } from '@/src/flows/ContractorOnboarding/components/ContractPreviewStatement';
+
 export type * from '@/src/flows/CostCalculator/types';
 
 export { useEmploymentQuery, useMagicLink } from '@/src/common/api';
@@ -97,6 +99,7 @@ export type {
   ButtonComponentProps,
   ZendeskDrawerComponentProps,
   DrawerComponentProps,
+  PDFPreviewComponentProps,
   Meta,
 } from '@/src/types/remoteFlows';